door closing completed
The door now closes seamlessly. All that's left is to reverse the process for when it opens. One note, making the variables public seems to create a separate variable. It treats it as a modifier of the original. If a varable has a value of 5 for x, the public varable treats that as default. To lesson the confusion, I think I'll stick to the variables in the editor.